requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=mkl
absl-py=0.11.0=pyhd3eb1b0_1
aiohttp=3.6.3=py38he774522_0
alabaster=0.7.12=pyhd3eb1b0_0
altair=4.1.0=pypi_0
appdirs=1.4.4=py_0
argh=0.26.2=py38_0
argon2-cffi=20.1.0=pypi_0
astor=0.8.1=pypi_0
astroid=2.5=py38haa95532_1
astunparse=1.6.3=py_0
async-timeout=3.0.1=py38haa95532_0
async_generator=1.10=pyhd3eb1b0_0
atomicwrites=1.4.0=py_0
attrs=20.3.0=pyhd3eb1b0_0
autopep8=1.5.5=pyhd3eb1b0_0
babel=2.9.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
base58=2.1.0=pypi_0
bcrypt=3.2.0=py38he774522_0
black=19.10b0=py_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
blinker=1.4=py38haa95532_0
blosc=1.21.0=h19a0ad4_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py38h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2020.10.14=0
cachetools=4.2.1=pyhd3eb1b0_0
certifi=2020.6.20=py38_0
cffi=1.14.5=py38hcd4344a_0
chardet=3.0.4=py38haa95532_1003
charls=2.1.0=h33f27b4_2
click=7.1.2=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyhd3eb1b0_0
coverage=5.4=py38h2bbff1b_2
cryptography=3.3.1=py38hcd4344a_1
cycler=0.10.0=py38_0
cython=0.29.22=py38hd77b12b_0
cytoolz=0.11.0=py38he774522_0
dask-core=2021.2.0=pyhd3eb1b0_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.6.0=pyhd3eb1b0_0
diff-match-patch=20200713=py_0
dlib=19.21.1=py38h082dfc2_0
docutils=0.16=py38_1
entrypoints=0.3=py38_0
findspark=1.3.0=py_1
flake8=3.8.4=py_0
flask=1.1.2=pypi_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py38_1
gast=0.4.0=py_0
giflib=5.2.1=h62dcd97_0
git=2.23.0=h6bb4b03_0
gitdb=4.0.5=pypi_0
gitpython=3.1.14=pypi_0
google-auth=1.27.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
google-pasta=0.2.0=py_0
graphviz=2.38=hfd603c8_2
grpcio=1.35.0=py38hc60d5dd_1
gunicorn=20.0.4=pypi_0
h5py=2.10.0=py38h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
imagecodecs=2021.1.11=py38h5da4933_1
imageio=2.9.0=py_0
imagesize=1.2.0=pyhd3eb1b0_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
intervaltree=3.1.0=py_0
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.20.0=py38hd4e2768_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pypi_0
isort=5.7.0=pyhd3eb1b0_0
itsdangerous=1.1.0=pypi_0
jedi=0.17.2=py38haa95532_1
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.1=pypi_0
joblibspark=0.3.0=pypi_0
jpeg=9d=h8ffe710_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.1=py38haa95532_0
jupyterlab-widgets=1.0.0=pypi_0
jupyterlab_pygments=0.1.2=py_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
keyring=22.0.1=py38haa95532_0
kiwisolver=1.3.1=py38hd77b12b_0
lazy-object-proxy=1.5.2=py38h2bbff1b_0
lcms2=2.11=hc51a39a_0
lerc=2.2.1=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libdeflate=1.7=h2bbff1b_5
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h33f27b4_0
libtiff=4.1.0=h56a325e_1
libzopfli=1.0.3=ha925a31_0
lz4-c=1.9.3=h2bbff1b_0
markdown=3.3.3=py38haa95532_0
markupsafe=1.1.1=py38he774522_0
matplotlib=3.3.4=py38haa244fe_0
matplotlib-base=3.3.4=py38h49ac443_0
mccabe=0.6.1=py38_1
mistune=0.8.4=py38he774522_1000
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.2.1=py38h46781fe_0
mkl_random=1.1.1=py38h47e9c7a_0
multidict=4.7.6=py38he774522_1
mypy_extensions=0.4.3=py38_0
nbclient=0.5.2=pyhd3eb1b0_0
nbconvert=6.0.7=py38_0
nbformat=5.1.2=pyhd3eb1b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5=py_0
notebook=6.2.0=pypi_0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
numpy-indexed=0.3.5=pypi_0
numpydoc=1.1.0=pyhd3eb1b0_1
oauthlib=3.1.0=py_0
olefile=0.46=py_0
opencv=4.0.1=py38h2a7c758_0
openjpeg=2.3.0=h5ec785f_1
openssl=1.1.1j=h2bbff1b_0
opt_einsum=3.1.0=py_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.1=py38hf11a4ad_0
pandoc=2.11=h9490d1a_0
pandocfilters=1.4.3=py38haa95532_1
paramiko=2.7.2=py_0
parso=0.7.0=py_0
pathspec=0.7.0=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.0=py38h4fa10fc_0
pip=20.3.3=py38haa95532_0
pluggy=0.13.1=py38_0
prometheus-client=0.9.0=pypi_0
prompt-toolkit=3.0.8=py_0
protobuf=3.14.0=py38hd77b12b_1
psutil=5.8.0=py38h2bbff1b_1
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=4.0.1=py38he44ac1e_0
py4j=0.10.9=pypi_0
pyarrow=3.0.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycodestyle=2.6.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pydeck=0.6.1=pypi_0
pydocstyle=5.1.1=py_0
pydot=1.4.2=pypi_0
pydotplus=2.0.2=py_3
pyflakes=2.2.0=pyhd3eb1b0_0
pygments=2.8.0=pyhd3eb1b0_0
pyjwt=1.7.1=py38_0
pylint=2.7.0=py38haa95532_1
pyls-black=0.4.6=hd3eb1b0_0
pyls-spyder=0.3.2=pyhd3eb1b0_0
pynacl=1.4.0=py38h62dcd97_1
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38ha925a31_4
pyreadline=2.1=py38_1
pyrsistent=0.17.3=py38he774522_0
pysocks=1.7.1=py38haa95532_0
pyspark=3.0.1=pypi_0
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=pyhd3eb1b0_0
python-decouple=3.4=pypi_0
python-jsonrpc-server=0.4.0=py_0
python-language-server=0.36.2=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py38he774522_2
pywin32=227=py38he774522_1
pywin32-ctypes=0.2.0=py38_1000
pywinpty=0.5.7=pypi_0
pyyaml=5.4.1=py38h2bbff1b_1
pyzmq=20.0.0=py38hd77b12b_1
qdarkstyle=2.8.1=py_0
qt=5.9.7=vc14h73c81de_0
qtawesome=1.0.1=py_0
qtconsole=5.0.2=pyhd3eb1b0_0
qtpy=1.9.0=py_0
regex=2020.11.13=py38h2bbff1b_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rope=0.18.0=py_0
rsa=4.7.1=pyhd3eb1b0_1
rtree=0.9.4=py38h21ff451_1
scikit-image=0.17.2=py38h1e1f486_0
scikit-learn=0.22.2.post1=pypi_0
scipy=1.6.1=py38h14eb087_0
send2trash=1.5.0=pypi_0
setuptools=52.0.0=py38haa95532_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=py38haa95532_0
smmap=3.0.5=pypi_0
snappy=1.1.8=h33f27b4_0
snowballstemmer=2.1.0=pyhd3eb1b0_0
sortedcontainers=2.3.0=pyhd3eb1b0_0
sphinx=3.5.1=pyhd3eb1b0_0
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.4=pyhd3eb1b0_0
spyder=4.2.1=py38haa95532_1
spyder-kernels=1.10.2=py38haa95532_0
sqlite=3.33.0=h2a8f88b_0
streamlit=0.78.0=pypi_0
tensorboard=2.3.0=pyh4dce500_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.3.0=mkl_py38h37f7ee5_0
tensorflow-base=2.3.0=eigen_py38h75a453f_0
tensorflow-estimator=2.3.0=pyheb71bc4_0
termcolor=1.1.0=py38haa95532_1
terminado=0.9.3=pypi_0
testpath=0.4.4=pyhd3eb1b0_0
textdistance=4.2.1=pyhd3eb1b0_0
three-merge=0.1.1=pyhd3eb1b0_0
tifffile=2021.1.14=pyhd3eb1b0_1
tk=8.6.10=he774522_0
toml=0.10.1=py_0
toolz=0.11.1=pyhd3eb1b0_0
tornado=6.1=py38h2bbff1b_0
traitlets=5.0.5=pyhd3eb1b0_0
typed-ast=1.4.2=py38h2bbff1b_1
typing_extensions=3.7.4.3=pyha847dfd_0
tzlocal=2.1=pypi_0
ujson=4.0.2=py38hd77b12b_0
urllib3=1.26.3=pyhd3eb1b0_0
validators=0.18.2=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=1.0.2=py38haa95532_1
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38_0
wrapt=1.12.1=py38he774522_1
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yapf=0.30.0=py_0
yarl=1.6.3=py38h2bbff1b_0
zeromq=4.3.3=ha925a31_3
zfp=0.5.5=hd77b12b_4
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
